% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_function.R
\name{obj_function}
\alias{obj_function}
\title{Objective function to be minimized}
\usage{
obj_function(pars, data, weights)
}
\arguments{
\item{pars}{Numeric vector of length 3}

\item{data}{Data frame}

\item{weights}{Function weights}
}
\value{

}
\description{
Objective function to be minimized
}
