% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for model optimization}
\usage{
prep_data(data, cum_cases = NULL, num_cases = NULL, date_var)
}
\arguments{
\item{data}{Data frame whihc must contains date information and at least one of cumulative cases or number of cases per date}

\item{cum_cases}{Character variable: name of cumulative cases column on data}

\item{num_cases}{Character variable: name of number of cases column on data}

\item{date_var}{Character variable: name date variable column on data}
}
\value{
A data frame with 4 variables:

\enumerate{
\item date
\item cum_cases
\item num_cases
\item d2: difference of num_cases between consecutive days}
}
\description{
Prepare data for model optimization
}
\examples{
prep_korea <- prep_data(data=korea_covid19, cum_cases="cum_cases", date_var="date")
str(prep_korea)
}
